Set JVM toolchain for Kotlin, remove configuration processor
