+++ goty-server/src/main/kotlin/com/aleinin/goty/game/GameSearchClient.kt
@@ -1,0 +1,5 @@
+package com.aleinin.goty.game
+
+interface GameSearchClient {
+    fun findGames(title: String, year: Int? = null, mainGame: Boolean, limit: Int): List<GameSearchResponse>
+}
